Combine both APIs into one, with choice of back-end as a parameter
This does not change any path on the API
